Rename detail modals
