--- app/src/views/private/components/modal-browse/index.ts
@@ -1,4 +1,0 @@
-import ModalBrowse from './modal-browse.vue';
-
-export { ModalBrowse };
-export default ModalBrowse;
