+++ app/src/views/private/components/modal-collection-detail/index.ts
@@ -1,0 +1,4 @@
+import ModalCollectionDetail from './modal-collection-detail.vue';
+
+export { ModalCollectionDetail };
+export default ModalCollectionDetail;
