+++ app/src/views/private/components/modal-collection-detail/modal-collection-detail.vue
