--- app/src/views/private/components/modal-detail/index.ts
@@ -1,4 +1,0 @@
-import ModalDetail from './modal-detail.vue';
-
-export { ModalDetail };
-export default ModalDetail;
