+++ app/src/views/private/components/modal-item-detail/index.ts
@@ -1,0 +1,4 @@
+import ModalItemDetail from './modal-item-detail.vue';
+
+export { ModalItemDetail };
+export default ModalItemDetail;
